Use taskfile variable to define primary Go module path
Although support was added to the Go module related "templates" for projects which contain modules at other paths, the
assumption remained that there will always be a primary module in the root of the repository, and this path was hardcoded
at several locations in the taskfiles.

It might be that a project that is not a Go module contains an accessory Go module in a subfolder. In this case, the use
of a taskfile variable will allow that path to be defined in one place, avoiding the need to override the hardcoded
incorrect location whenever a Go module-related task is run manually.