Use more common Cobra command root function in "docsgen" template
The function that creates the Cobra CLI command root is declared by each project, and so might have any name. As it was
the most advanced in general, I used Arduino Lint as the source for the Cobra-based project website template assets.
However, I noticed that Arduino CLI and Arduino Firmware Uploader both use the function name `cli.NewCommand()`, so I
think this is more likely to be the name used by tooling projects this template is applied to.