Document project-specific command root generation function
The function that creates the Cobra CLI command root is declared by each project, and so might have any name. The
"docsgen" program that generates the command line reference content for the website calls this function, so it will be
necessary for the appropriate function to be added to the program when it is added to a project.

Previously, there was no documentation of this project-specific component of the "template", which could result in
confusion when it was applied to a project that happens to have a different function name from the `cli.NewCommand()`
used by the "template" docsgen code.