Use explicit package name in "docsgen" CLI package import
The template "docsgen" program for generating the website Cobra CLI reference content relies on the project's Cobra CLI
package being named `cli`. But this is not guaranteed to be so, and in fact the Cobra documentation uses `cmd` instead.

Confusion caused by applying the program to projects that use different package names will be avoided by specifying an
explicit package name in the import statement.