Use shallow fetch for Go tester build
The default behavior of the `actions/checkout` action is to do a shallow fetch of the repository, which is the most
efficient if all that's needed is a copy of the repository files. In cases where the repository history is needed, this
behavior is not appropriate, and so it can be configured via the `fetch-depth` input. Setting this input to 0 causes the
full history to be fetched.

A full fetch is required for workflows that use the `arduino/create-changelog` action to generate a raw changelog from
the commit history, and so it is found in the "Release" workflow. However, there is no use of the commit history by the
"Publish Tester Build" workflow, which does not need a changelog.

The unnecessary full fetch makes the workflow less efficient and more difficult to understand, so it must be removed.