Sort tasks in alphabetical order
I originally set out to establish some logical order to the tasks in the taskfiles I've worked in. Based from the start
on some indecipherable and subjective criteria, then randomized to some extent by the whims of Git over the course of
development, it ends up a jumble.

Alphabetical order is completely objective and it results in a fairly logical order in the end due to the practice of
using prefixes on the task names according to their domain. The exception I make is placing the convenience "umbrella"
tasks at the top (in their own alphabetical order) so that they will have the maximum visibility as the most useful tasks.
Even that is not completely inconsistent, since these don't have a namespace prefix, and so might be imagined to have a
null global namespace prefix that should be sorted first.