Add minimal GitHub release creation template workflow
The requirement emerged for a release workflow that can be applied to projects that don't have any release assets to
build and upload.

On every push of a release tag (e.g., v1.2.3) to the repository, create a GitHub release with a raw changelog generated
from the commit history since the last release tag.

Tags that contain a pre-release version will cause the GitHub release to be marked as a pre-release.