Add explanatory comments for `LDFLAGS` taskfile variable
This variable has a specific usage in the release-go-task template for injecting versioning information during the
builds. The `go:build` task is intended for use by all Go application projects, even if the release system is not
applicable. For this reason, the "go-task" template taskfile contains an empty `LDFLAGS` taskfile variable definition,
which may be defined with any `-ldflags` flag the project might need for its build, or left empty if the project does
not have any need for such a flag.

The usage of this variable might not be clear to the template user, so a "TODO" comment is added, which will be removed
during the installation of the "template" in a project. In the event the variable it is left empty, its presence in the
taskfile might be the source of confusion, and so a permanent explanatory comment is also provided.