Make "dependencies" an alias for "topic: infrastructure" label
By default, dependabot adds the "dependencies" repository label to its pull requests, but . The labeling behavior can be
customized in the `dependabot.yml` configuration file, but this configuration file is for general outdated dependency
updates, whereas many projects use dependabot only for security updates, which are handled by the repository settings
rather than the configuration file. It does not appear that there is any way to adjust the label name via the repository
settings.