Improve wording of "topic: infrastructure" label description
The previous description of the "topic: infrastructure" repository label was:

> Related to repository infrastructure

This might give the impression that it is intended to be only for Git-related things (e.g., `.gitignore`), but actually
this is for anything related to the infrastructure of the project (e.g., build system, dependencies) which includes, but
is not limited to, Git-related things.

A more general description will more effectively communicate the intended meaning of the label.