Add shared phase labels for tooling projects
Two new labels are added to the shared repository label configuration file for tooling projects. These will allow the
current stage in the development work to resolve an issue to be indicated.

At the moment, the phases recognized by the tooling team are design and implementation. Additional labels will be added
if more distinct phases emerge in practice.