Add criticality labels to "universal" shared label configuration file
A set of "criticality" labels allows the assignment of an impact level estimation to issues and pull requests.

The usage of these labels is distinct from the priority labels in that priority is based on multiple factors. In addition
to impact, the effort required to resolve must also be considered when assigning priority.