Clarify that `sh` is not required in template install docs
A significant portion of Arduino users use Windows. The first option listed in the "template" installation instructions
is for an installation shell script, which can be used on Window, but might not be convenient for some Windows users.
This could be discouraging or frustrating to potential new users, but not necessarily so because the script is an
optional convenience and directly downloading and installing the tool is also an option, as documented in the
instructions for the second listed option.

A little refinement of the wording of the script's `sh` requirement makes the situation clear to those users.