Add missing reference link re: `PATH` to template install docs
A reference link provides easy access to the relevant information for readers who are not familiar with this concept
while not harming readability for everyone else by increasing verbosity.

This had already been done for the script method installation instructions, but not for the direct download method
instructions. Since the user is only intended to read the instructions for the specific installation method they choose,
the reference link must be added to the first mention of `PATH` in each set of instructions.