Pin Windows-compatible version of Task where necessary
The recent 3.9.1 release of the Task task runner does not correctly handle the PATH environment variable on Windows.
Since the "template" workflows were configured to always use the latest version of Task within the 3.x major version
series, this bug caused spurious failure of workflows under the following conditions:

- Used a Windows runner
- Ran a task that contained a command reliant on a file from PATH

The workaround is to pin the version of Task used in these workflows to the last working version: 3.9.0.