Add tests for PATH suggestion output of installation script
The "template" installation script stored in this repository prints advice to the user about adding the installation to
their PATH environment variable.

This advice is adjusted according to whether a previous installation was found in the PATH and according to the path(s)
of their installation(s) so it is fairly complex. For this reason, it will be good to have a check in the
"Test install script" CI workflow to ensure the output is as expected.