Fix install script's check for previous installation
The "template" installation script stored in this repository checks for an existing installation in the PATH in order to
provide appropriate advice to the user about adding the installation to their their PATH environment variable.

This check is done using `command -v`. It turns out that the exit status is shell dependent in the event the command is
not found, so that it might be either 1 or 127 depending on the user's system. The script previously assumed that the
exit status would be 1 when the command was not found in PATH, which resulted in spurious advice under these conditions:

```
An existing arduino-lint was found at . Please prepend "/home/runner/work/tooling-project-assets/tooling-project-assets/bin" to your $PATH or remove the existing one.
```

It seems safest to invert the logic so that the advice about an existing installation in PATH is only printed when one
was found.