Check exit status directly in install script
The "template" installation script stored in this repository checks for an existing installation in the PATH in order to
provide appropriate advice to the user about adding the installation to their their PATH environment variable.

This check is done by checking the exit status of `command -v`. The previous use of `$?` for that check was unnecessary
since the command itself can be checked.

This change is required for compliance with ShellCheck rule SC2181:

https://github.com/koalaman/shellcheck/wiki/SC2181