Add `release-go-crosscompile-task` workflow that use golang crosscompile
Previously there was only a workflow in the assets repo handling the release of a go project. That workflow was based on the build process we have in the Arduino CLI, but that's not the only way of compiling. We use that one in the Arduino CLI because we need CGO enabled and is not possible with golang crosscompile. The easiest (and fastest) way of cross-compiling is by using environment variables to set the OS and ARCH we want to build for. See https://go.dev/doc/install/source#introduction