Remove aliases from shared label configurations
The "GitHub Label Sync" tool used by the "Manage Labels" template workflow has a bug that can cause unwanted label
deletion when multiple labels in the repository match an element in the configuration file.

For example, the configuration for the universally shared "type: enhancement" label defines the "enhancement" label as
an alias, so if a repository has both a "type: enhancement" and an "enhancement" label (which might happen through a
maintainer who isn't aware of the label management system adding a label manually), both labels will match that
configuration element. The bug would cause the deletion of the "type: enhancement" label from the repository, with it
the irreversible loss of labeling in all issues and pull requests it had been applied to.

This multiple match condition can only occur when aliases have been configured in the element. So the mitigation for the
bug is to remove all alias configurations.

Because these aliases are very useful for automatically migrating a repository from the existing use of GitHub's
inadequate stock labels to our standardized labels after the initial installation of the "Manage Labels" workflow in a
repository (or even when running the tool directly in case a maintainer prefers to manage their labels manually), this
provisional change will be reverted as soon as the bug has been fixed in "GitHub Label Sync" and released.