Remove `notes` from allowed properties in label configuration schema
A JSON schema is used to validate the shared repository label configuration files hosted in this repository in order to ensure they are:

- Compliant with the restrictions of the GitHub labels API
- Compliant with the data format of the "GitHub Label Sync" tool used to manage the labels
- Consistent with the standardized labels format used in Arduino tooling projects

Validation of the configuration file was added to the "GitHub Label Sync" tool in the recent 2.1.0 release. Before
2.1.0, the tool ignored any additional properties in the label configuration objects. It now errors if there are any
unexpected properties.

At the time the original configuration file was developed, it was in JSON, which does not support comments. I found the
need to add some internal explanatory commentary to some of the labels, so I added an arbitrary "notes" key as a
container for this information, and our JSON schema was also configured to accept this field.

This "notes" field now causes the configuration files to be considered invalid by the tool, so our schema must do the
same.