Apply base name schema to `aliases` fields of label configuration file
A JSON schema is used to validate the shared repository label configuration files hosted in this repository in order to ensure they are:

- Compliant with the restrictions of the GitHub labels API
- Compliant with the data format of the "GitHub Label Sync" tool used to manage the labels
- Consistent with the standardized labels format used in Arduino tooling projects

The "GitHub Label Sync" tool supports an "aliases" array in the label configuration, which defines the labels which
should be renamed according to the configuration. This array is a list of label names, so the same fundamental
restrictions (both from GitHub and "GitHub Label Sync", but not Arduino's style guidelines) apply to it as to the `name`
field.