Remove `notes` field from label configuration files
At the time the original configuration file was developed, it was in JSON, which does not support comments. I found the
need to add some internal explanatory commentary to some of the labels, so I added an arbitrary `notes` key as a
container for this information, and our JSON schema was also configured to accept this field.

I later decided to convert the files to YAML, since that is the language used by the majority of the asset configuration
files. At this point it became possible to use comments, but the `notes` field was already in place so it
seemed pointless to change it.

Validation of the configuration file was added to the "GitHub Label Sync" tool in the recent 2.1.0 release. Before
2.1.0, the tool ignored any additional properties in the label configuration objects. It now errors if there are any
unexpected properties.

This `notes` field now causes the configuration files that contain it to be considered invalid by the tool:

```
$ github-label-sync --labels https://raw.githubusercontent.com/arduino/tooling-project-assets/main/workflow-templates/assets/sync-labels/universal.yml per1234/some-repo
Syncing labels for "per1234/some-repo"
Validating provided labels
Invalid label:
  {"name":"conclusion: moved","color":"940404","description":"Issue/PR moved to another repository","notes":"For use when it's not possible to do a true transfer\n(PRs, different organizations, or maintainer doesn't have write permissions in target repo).\n"}
  - must NOT have additional properties

[...]
```

So the `notes` field is hereby removed, with the contents moved to comments.