Update `actions/setup-python` reference link to match version in use
The reference link describes the version formats supported by the `actions/setup-python` GitHub Actions action used to
setup Python in the GitHub Actions runner for the workflows. In order to provide the correct documentation for the
action version in use, the URL is based on the major version ref of the action repo, but this was not updated at the
time of the bump from the v2 to v3 version series in the workflows.