Bump semver-tool to 3.2.0
This was the latest version at the time the "Release Tag" workflow was written.

These other workflows were using an older version. The 3.2.0 is already outdated, and ideally all the workflows would be
updated, but this is the version has been well tested and used for some time in the "Release Tag" workflow. So this is
intended to at least bring all the workflows into sync.