Standardize format of placeholders in templates
The assets collected in this repository are designed to be as universally applicable as is practical.

However, in some cases it is necessary to make some project-specific configurations within the asset files during the
installation. These locations are indicated by "TODO"-prefixed comments where possible and the location where the
configuration value is to be filled is indicated by placeholder text. I have found that it is sometimes easy to miss
placeholders among a sea of visually similar variable names both while installing and while reviewing installations made
by others.

Adding a `TODO_` prefix to the descriptive suffix of the placeholder text will make it easy for the installer to find
all places where configuration is needed by a quick text search for "TODO" (which will now find both the comments and
placeholders) and also make it easy for the reviewer to spot placeholders the installer missed.