Split configuration file validation infrastructure by type
Previously, the validation of various configuration files stored and/or used in the repository was done by a single
monolithic task and workflow. Due to the different characteristics of the files and their JSON schemas, there was no
real benefit from this approach in terms of efficiency or maintainability.

As validation of additional configuration file types became necessary, the previous system grew more unwieldy. Splitting
it up into separate tasks and workflows offers several primary benefits:

- Tasks are more maintainable
- CI workflows are more efficient due to only running validations relevant to the specific configuration files that were
  modified