Move docs for issue template chooser configs alongside assets
A collection of reusable GitHub issue template chooser configuration files are stored in this repository.

Previously, their documentation was located in the documentation for the issue templates. The reasoning for that
approach was that the template chooser feature is exclusive to repositories that have issue templates. However, it does
not scale well as additional issue templates are added, and also makes it difficult to find information about these
template chooser configuration files.

A better approach is to place the documentation alongside the assets and then reference that from associated assets.