Add a collection of reusable issue form configuration files
High quality feedback via GitHub issues is a very valuable contribution to the project. It is important to make the
issue creation and management process as efficient as possible for the contributors, maintainers, and developers.

Issue templates are helpful to the maintainers and developers because it establishes a standardized framework for the
issues and encourages the contributors to provide the essential information.

When a repository is configured to use the GitHub issue forms system, the contributor is now presented with a web form
when creating an issue. This consists of multi-line input fields that have the same formatting and attachment
capabilities as the standard GitHub Issue composer, in addition to menus and checkboxes where appropriate.

The use of this form-based system should provide a much better experience for the contributors and also result in higher
quality issues by establishing a standardized framework for the issues and encouraging contributors to provide the
essential information.

The forms are configured via a YAML file. A collection of reusable issue form configuration files designed for each
project type accompanied by documentation for their use are now stored in this repository. These will enable easily
setting up issue forms in any project repository and efficiently maintaining the shared files in a centralized location.

On every push and pull request which modifies relevant files, and periodically, the issue form configuration files will
be validated against their JSON schema to catch any problems with the data.