Add a version of the "Sync Labels" workflow for npm-managed projects
We are now standardizing on the use of npm to manage tool dependencies of the assets.

There is one drawback to that approach, which is that it requires the addition of the npm metadata files to the
repository (if not already present).

Two more files are not of any concern for the average Arduino tooling project. However, the "Sync Labels" workflow is
already in widespread usage in other types of official Arduino projects. In some projects (e.g., Arduino libraries),
where the contents are intended to be very approachable to less advanced users, the added files may be considered
problematic, especially since, unlike all the other workflows that were candidates for npm-managed tools, the
"Sync Labels" workflow did not require the addition of any files other than the GitHub Actions workflow (the others also
require a taskfile at minimum). For this reason, the previous lightweight version of "Sync Labels" is retained, but
another version added for use in projects where the dependency management is desired.