Use a global workflow variable to define the "Draft 4" compatible version of ajv-cli
The ajv-cli tool is used for validation of data files against their JSON schema.

In general, it is preferable, and for some schemas even mandatory, to use the modern versions of ajv-cli. However,
support for the "Draft 4" schema specification was dropped in ajv-cli version 4.0.0. So when working with JSON schemas
that specify that draft, it is necessary to use ajv-cli 3.3.0, the last compatible version.

Such schemas are relatively common, likely partly due to the Schema Store's policy that "Draft 4" be used whenever
possible. Having this version defined in a single place makes taskfiles with multiple tasks subject to this somewhat
esoteric (and thus requiring an explanatory comment) information a bit cleaner.