Host standardized ClangFormat configuration file
ClangFormat is now the preferred code formatting tool for use in tooling and Arduino language/C++/C firmware projects.

An Arduino code style has been established by the Arduino IDE's "**Auto Format**" feature and the code of the sketches
and libraries written by Arduino. It is important that this style remain consistent even through this time of transition
from the former preferred Artistic Style code formatter tool. A ClangFormat configuration was carefully developed to
achieve this goal.

This complex configuration must be properly maintained through updates of the actively developed LLVM/clang/ClangFormat
project. This can only be achieved if there is a single canonical source of the file, from which any enhancements and
fixes are pulled from and pushed to by the various projects which use it, such as:

- Arduino IDE 2.x
- Arduino Language Server
- Firmware projects (via the "Check C++" workflow template)
- Arduino Cloud (pending)

This repository will now be used to host and maintain that canonical upstream copy.

This is the unmodified configuration which is already in use by Arduino IDE 2.x and Arduino Language Server.

The YAML format source:

https://github.com/arduino/arduino-language-server/blob/0.7.1/ls/ls_formatter.go#L15-L160

Which is functionally equivalent to the JavaScript object generated from it now in use by Arduino IDE 2.x:

https://github.com/arduino/arduino-ide/blob/2.0.0-rc9/arduino-ide-extension/src/node/clang-formatter.ts#L133-L278