Validate ClangFormat configuration file against its JSON schema
A `clang-format:validate` task is added to the repository infrastructure. In addition to the option to run this task
locally during maintenance on the file, it is executed automatically by the repository's CI system on every change to
relevant files and periodically to catch problems exposed by external changes (e.g., schema updates).