Add assets for JavaScript code linting
An ESLint configuration file provides the standardized configuration for use in Arduino tooling projects. This uses the
Airbnb base configuration. Since Prettier provides code formatting, overlapping ESLint rules are disable via
eslint-config-prettier.

A task is provided to run ESLint on the JavaScript code files of the project the assets are applied to.

On every push and pull request that affects relevant files, and periodically, a GitHub Actions workflow runs this task.