Add script to convert ClangFormat configuration file for use in Arduino IDE 2.x
The Arduino IDE 2.x "Auto Format" feature uses ClangFormat to format the sketch file open in the editor.

The default ClangFormat configuration is embedded in object form in the Arduino IDE code base, not in the YAML format
used by `clang-format`.

A script makes the required conversion.

The `clang-format:convert` task executes the script.

In order to facilitate updates, A GitHub Actions workflow runs the conversion on every change to relevant files and
makes it available for download as a workflow artifact.