Format ClangFormat configuration goldens according to traditional Arduino code style
The base level Arduino code style is provided by the default output of the Arduino IDE 1.x "Auto Format" feature.

The Arduino IDE 2.x Auto Format feature must be configured to reproduce that style as closely as possible. For this
reason, the "golden master" test data is formatted to that style. This will be compared against the output generated by
Arduino's ClangFormat configuration.

The official Arduino sketches used as "samples" test data was imported into the repository exactly as it exists in the
upstream projects. Unfortunately, some of that code does not comply with the base Arduino code style, so it must be
brought into compliance by formatting using the 2.05.1 version of the Artistic Style code formatting tool used by Arduino IDE 1.8.19:

https://github.com/arduino/Arduino/blob/1.8.19/build/build.xml#L483

with the default formatter configuration of Arduino IDE 1.8.19:

https://github.com/arduino/Arduino/blob/1.8.19/build/shared/lib/formatter.conf