Format ClangFormat configuration goldens using ClangFormat 11.0.1
The ClangFormat configuration file hosted in this repository was developed for ClangFormat 11.0.1, the version in use by
Arduino Language Server and Arduino IDE 2.x at that time.

Significant effort was made to create a configuration that would result in the same output as that of the Arduino IDE
1.x "Auto Format" feature. Where differences between ClangFormat and the Artistic Style formatter tool of Arduino IDE
1.x made that impossible, the configuration was based on surveys of existing official Arduino code.

This commit demonstrates the unavoidable changes that resulted from the configuration tool change.