Add targeted ClangFormat configuration test data
These files specifically target each relevant configuration setting of ClangFormat 14.0.0, the version currently in use
by Arduino IDE 2.x and Arduino Language Server.

In combination with the real world Arduino sketch samples, this should provide full coverage of the result of proposed
changes to the configuration file or updates to the ClangFormat version in use.

The "golden masters" were formatted using ClangFormat 11.0.1, which is the version the configuration file was developed
for.