Compare results of ClangFormat configuration file against those expected
ClangFormat is a very complex tool under active development. This means that unintended changes to the formatter output
might result from:

- Unintended side effects of intentional configuration changes.
- Differences in ClangFormat's handling of existing configuration options after updating to a new version.
- Introduction of new configuration options, which will be set to the LLVM style default value if not set in Arduino's
  configuration, after updating to a new version.

For this reason, it is necessary to carefully validate the configuration in preparation for:

- Changes to the configuration file
- Updating the version of ClangFormat used in Arduino's projects

Validation is done by formatting a collection of test data files and then comparing the result against "golden masters"
which have the intended formatting. The repository's continuous integration system does this check after every change to
relevant files. It can also be run locally using the `clang-format:check-configuration` task.

Arbitrary ClangFormat versions may be specified for use in the check in order to evaluate a version of ClangFormat under
evaluation for an update candidate.