Move notes out of ClangFormat configuration to dedicated file
The ClangFormat configuration file hosted in this repository is used by several projects.

Previously, comments in the configuration file were used to record some potentially useful information about some of the
configuration options. This non-functional content increases the burden of maintaining the downstream copies. A more
efficient approach is to have a single copy of the information stored here, easily accessible from the downstreams via
the source link.