Standardize ClangFormat configuration file format
The LLVM/Clang/ClangFormat project is under very active development, with configuration keys regularly being added,
changed to a different data type, or removed.

When a key is set in the configuration file, the default LLVM style value is used. This means that updating to a new
version of ClangFormat might result in unintended changes to the results from the configuration file. In the case of a
change of type, it seems the ClangFormat developers generally will provide backwards compatibility of some sort, but
there is no guarantee that the resulting configuration will be exactly the same as was intended when the value of the
key was selected for use in Arduino's configuration file.

ClangFormat has a useful `--dump-config` flag which outputs the effective configuration. This allows us to determine
whether there are any differences between the intended configuration as defined by Arduino's configuration file and the
actual configuration of the tool.

In order to accomplish this, it is best to use the exact format output by `clang-format --dump-config`, which allows any
differences to be detected and seen clearly using a simple diff command.

The only exception is the insertion of a comment providing the source URL of the configuration file in order to
facilitate the syncing of changes with the downstream copies of the file in the various projects that use it.